[spark] Fix analyze compatibility with spark3.3- (#2835)
